Create a test for getStatusBoard method and change the types of the board variables

+++ src/main/java/it/unibo/puzbob/model/Board.java
@@ -10,7 +10,7 @@ public interface Board {
 
     Ball[][] getStatusBoard(); 
     
-    Pair<Integer, Integer> getBoardSize();
+    Pair<Double, Double> getBoardSize();
 
     ArrayList<String> getColors();
 }
